+++ packages/perseus/src/widget-ai-utils/cs-program/cs-program-ai-utils.ts
