--- packages/perseus/src/widget-ai-utils/cs-program/prompt-utils.test.ts
@@ -1,13 +1,0 @@
-import {getPromptJSON} from "./prompt-utils";
-
-describe("CS Program getPromptJSON", () => {
-    it("it returns JSON with the expected format and fields", () => {
-        const resultJSON = getPromptJSON();
-
-        expect(resultJSON).toEqual({
-            type: "cs-program",
-            isSupported: false,
-            message: "",
-        });
-    });
-});
